# api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : " {{ iFind, the Lost and Found WebApp }}",
    "version" : "1.0.0",
    "title" : "iFind-API",
    "contact" : {
      "email" : "nignanthomas@gmail.com"
    }
  },
  "host" : "localhost:5000",
  "basePath" : "/api/v1/",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/category" : {
      "post" : {
        "tags" : [ "admin" ],
        "summary" : "Create a category",
        "description" : "Create category",
        "operationId" : "Addcategory",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "header",
          "name" : "Authorization",
          "description" : "A token header",
          "required" : true,
          "type": "string"
        },
        {
          "in" : "body",
          "name" : "body",
          "description" : "Category object to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Category Created"
          },
          "401" : {
            "description" : "Unathorized - Signin to have access"
          },
          "409" : {
            "description" : "Category Already Exists"
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      },
      "get" : {
        "tags" : [ "admin" ],
        "summary" : "Get all categories",
        "description" : "All categories",
        "operationId" : "allCategories",
        "parameters" : [],
        "responses" : {
          "200" : {
            "description" : "Categories fetched successfully"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Category" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "Electronics"
        },
        "description" : {
          "type" : "string",
          "example" : "Phones, Computers, Earphones, etc"
        }
      },
      "xml" : {
        "name" : "Category"
      }
    }
  }
}